requirements.txt
pywal ~= 3.3.0
pyyaml ~= 6.0
